break-out coherence-function.ts from metamodel.ts
