+++ lib/js/metamodel/metamodel.ts
@@ -47,6 +47,13 @@ export {
     SERIALIZE_FORMAT_OBJECT
 }
 
+import {
+    CoherenceFunction
+} from './coherence-function.ts';
+
+export {
+    CoherenceFunction
+}
 
 /**
  * Building blocks and helper functions to define a Model.
@@ -55,47 +62,6 @@ export {
 const _NOTDEF = Symbol("_NOTDEF");
 type NotDef = typeof _NOTDEF;
 export type DefaultProvided<D> = Exclude<D, NotDef>;
-
-// FIXME/TODO: type can be anything that _AbstractStruct.get would return
-// Also as Proxies!
-type ValueMap = Record<string, unknown>;
-type CoherenceFn = (valueMap: ValueMap) => void;
-
-export class CoherenceFunction {
-    // Use the definite assignment assertion '!'
-    // This tells TypeScript the property WILL be initialized in the constructor.
-    public readonly name!: string;
-    public readonly fn!: CoherenceFn;
-    public readonly dependencies!: FreezableSet<string>;
-    constructor(dependencies: string[], fn: CoherenceFn /*(valueMap)*/) {
-        Object.defineProperties(this, {
-            dependencies: {
-                value: Object.freeze(new FreezableSet(dependencies)),
-            },
-            fn: { value: fn },
-            name: { value: fn.name || "(anonymous)" },
-        });
-        Object.freeze(this);
-    }
-
-    static create(
-        dependencies: string[],
-        fn: CoherenceFn,
-    ): [string, CoherenceFunction] {
-        const instance = new this(dependencies, fn);
-        return [instance.name, instance];
-    }
-
-    // This way it goes nicely into the struct definition without
-    // having to repeat the name!
-    get nameItem(): [string, CoherenceFunction] {
-        return [this.name, this];
-    }
-
-    toString() {
-        return `[CoherenceFunction ${this.name}]`;
-    }
-}
 
 export function sort_alpha(a: string, b: string): number {
     return a.localeCompare(b, undefined, { sensitivity: "base" });
